Last Commit

+++ src/main/java/org/example/appDonovan/entity/EntityInterface.java
@@ -1,6 +1,6 @@
 package org.example.appDonovan.entity;
 
 public interface EntityInterface {
-    Long getId();
-    void setId(Long j);
+    Integer getId();
+    void setId(Integer j);
 }
